# specs/001-podcast-summary-ui/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Podcast Summary API
  version: 0.1.0
paths:
  /api/podcasts/ingest:
    post:
      summary: Submit a podcast URL for ingestion (use existing transcript if present, otherwise transcribe and summarize)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
              required: [url]
      responses:
        "202":
          description: Ingestion job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAccepted"
        "400":
          description: Invalid or unreachable URL
        "409":
          description: Job already in progress for this URL
...
components:
  schemas:
    JobAccepted:
      type: object
      properties:
        jobId:
          type: string
        podcastId:
          type: string
        status:
          type: string
          enum: [queued, running, succeeded, failed]
      required: [jobId, podcastId, status]
...